tech: cleanup frontend routes and files
