+++ .gitlab-ci/templates/.env.j2
@@ -14,5 +14,3 @@ MONITORENV_HOST={{ rapportnav_monitorenv_host }}
 
 SENTRY_DSN={{ rapportnav_sentry_dsn }}
 SENTRY_SECU_HEADER_ENDPOINT={{ rapportnav_sentry_secu_header_endpoint }}
-
-REACT_V2_ACTIVATED=true
