+++ frontend/.env.example
@@ -3,4 +3,3 @@
 # Sentry
 
 FRONTEND_SENTRY_DSN=https://something@sentry.incubateur.net/121
-REACT_V2_ACTIVATED=true
