+++ frontend/.env.local.defaults
@@ -1,5 +1,4 @@
 ################################################################################
 # Sentry
 
-REACT_V2_ACTIVATED=true
 FRONTEND_SENTRY_DSN=https://8857258f9f1549968b13e15759bdf2bc@sentry.incubateur.net/121
