--- frontend/src/features/common/utils/control-utils.test.ts
@@ -1,5 +1,0 @@
-describe('control utils', () => {
-  it('dummy', () => {
-    expect(false).toBe(false)
-  })
-})
