+++ frontend/src/v2/pages/__tests__/signup-page.test.tsx
@@ -1,0 +1,9 @@
+import { render, screen } from '../../../test-utils.tsx'
+import SignupPage from '../signup-page.tsx'
+
+describe('SignupPage', () => {
+  test('should render', () => {
+    render(<SignupPage />)
+    expect(screen.getByText("S'inscrire")).toBeInTheDocument()
+  })
+})
