+++ frontend/src/v2/pages/home-page.tsx
@@ -2,7 +2,7 @@ import { useGlobalRoutes } from '@router/use-global-routes'
 import { FC, useEffect } from 'react'
 import { useNavigate } from 'react-router-dom'
 
-const Home2: FC = () => {
+const HomePage: FC = () => {
   const navigate = useNavigate()
   const { homeUrl } = useGlobalRoutes()
 
@@ -14,4 +14,4 @@ const Home2: FC = () => {
   return <div />
 }
 
-export default Home2
+export default HomePage
