+++ frontend/src/v2/pages/login-page.tsx
@@ -1,12 +1,12 @@
 import { FC } from 'react'
 import { useNavigate } from 'react-router-dom'
-import PageWrapper from '../features/common/components/ui/page-wrapper.tsx'
-import AuthToken from '../features/auth/utils/token.ts'
+import PageWrapper from '@common/components/ui/page-wrapper.tsx'
+import AuthToken from '@features/auth/utils/token.ts'
 import { FlexboxGrid, Stack } from 'rsuite'
 import { Accent, Button, FormikTextInput, Size } from '@mtes-mct/monitor-ui'
 import { Form, Formik, FormikHelpers } from 'formik'
 import { validate } from 'email-validator'
-import { csrfToken } from '../features/auth/utils/csrf.ts'
+import { csrfToken } from '@features/auth/utils/csrf.ts'
 import { trim } from 'lodash'
 
 export const LOGIN_ENDPOINT = '/api/v1/auth/login'
@@ -27,7 +27,7 @@ const initialValues: LoginFormValues = {
 
 const authToken = new AuthToken()
 
-const Login: FC = () => {
+const LoginPage: FC = () => {
   const navigate = useNavigate()
 
   const handleSubmit = async (
@@ -148,4 +148,4 @@ const Login: FC = () => {
   )
 }
 
-export default Login
+export default LoginPage
